added the modelrunner script
The script takes care of running the models that are queued. it will
work as a cronjob. So the script needs to be added to crontab. It get
the database config of the app and get all the queued model runs and
creates a process for each modelrun. While adding to crontab it needs to
be configured with virtualenv that its running in