Refactored the app for the newer approach
Changes:
* Config will be  set through the environment variables with prefix
VWADAPTOR_
* resources will no longer be saved locally, instead it will use a cloud
provider to store the actual files. for dev it will mimic localstorage
* Resource will no longer have the resource_url field. it will rather be
dynamically generated to be downloaded from the storage url. it will
only contain the resource_name attribute which is same as the storage
object.
* For storage a modified version of flask-cloudy is used which uses
apache-libcloud under the hood.
* To make swift work with flask-cloudy had to make a little change in
the apache-libclouds storage driver. an apache libcloud fork is used
instead of using the release version.
* made changes in modelrun.start method. it now checks if all the
resources for the modelruns has been uploaded before queuing the model.